+++ include/axon/hdfs.h
@@ -42,17 +42,17 @@ namespace axon
 			bool chwd(std::string);
 			std::string pwd();
 			bool mkdir(std::string);
-			int list(const axon::transfer::cb &);
-			int list(std::vector<entry> &);
-			long long copy(std::string, std::string, bool);
-			long long copy(std::string src, std::string dest) { return copy(src, dest, false); };
+			size_t list(const axon::transfer::cb &);
+			size_t list(std::vector<entry> &);
+			off_t copy(std::string, std::string, bool);
+			off_t copy(std::string src, std::string dest) { return copy(src, dest, false); };
 			bool ren(std::string, std::string);
 			bool del(std::string);
 
 			int cb(const struct entry *);
 
-			long long get(std::string, std::string, bool);
-			long long put(std::string, std::string, bool);
+			off_t get(std::string, std::string, bool);
+			off_t put(std::string, std::string, bool);
 
 			bool set(char, std::string);
 			bool set(char, int);
